code/research/routers.py
from django.urls import path, include
from rest_framework import routers, serializers, viewsets

from research.views.uf_view import UFViewSet

# Routers provide an easy way of automatically determining the URL conf.
router = routers.DefaultRouter()
router.register(r'uf', UFViewSet)

# Wire up our API using automatic URL routing.
# Additionally, we include login URLs for the browsable API.
urlpatterns = [
    path('', include(router.urls)),
]