code/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
django-filter  # Filtering support
#browsable API
markdown
pyyaml 
uritemplate
#db
psycopg2-binary==2.9.4
